+++ src/trx/game/objects/vehicles/mounted_gun.h
@@ -1,0 +1,3 @@
+#pragma once
+
+bool MountedGun_Control(void);